+++ test/OpenTelemetry.Tests/Metrics/MetricTestData.cs
@@ -40,8 +40,7 @@ internal static class MetricTestData
     public static TheoryData<double[], HistogramConfiguration, double, double> ValidHistogramMinMax => new()
     {
         { [-10.0, 0.0, 1.0, 9.0, 10.0, 11.0, 19.0], new HistogramConfiguration(), -10.0, 19.0 },
-        { [double.NegativeInfinity], new HistogramConfiguration(), double.NegativeInfinity, double.NegativeInfinity },
-        { [double.NegativeInfinity, 0.0, double.PositiveInfinity], new HistogramConfiguration(), double.NegativeInfinity, double.PositiveInfinity },
+        { [double.NegativeInfinity, 0.0, double.PositiveInfinity], new HistogramConfiguration(), 0.0, 0.0 },
         { [1.0], new HistogramConfiguration(), 1.0, 1.0 },
         { [5.0, 100.0, 4.0, 101.0, -2.0, 97.0], new ExplicitBucketHistogramConfiguration { Boundaries = [10.0, 20.0] }, -2.0, 101.0 },
         { [5.0, 100.0, 4.0, 101.0, -2.0, 97.0], new Base2ExponentialBucketHistogramConfiguration(), 4.0, 101.0 },